requirements.txt
streamlit==1.32.0
edge-tts
faster-whisper
moviepy==1.0.3
pandas
pyyaml==6.0.1
google-generativeai==0.3.2
openai==1.12.0
ollama==0.1.6
pillow==10.2.0
numpy==1.26.4
tqdm==4.66.2
requests==2.31.0
replicate==0.22.0
firebase-admin==6.4.0
python-dotenv==1.0.1 